extends slf4jmedia to override name in mdc
until this change the name in the mdc was the same as the name of the property
e.g. the property was name the mdcname was name.
now it is possible to override this by providing a map instead of a list.
Map.of("name", "username") and the property name will get username as it's mdcname