replace writer with outputstream in writebale
this change should allow to use also binary media to be use this interface.
to create a string of a writebale it has a asString default method now.
this means instead of
  String str = writable.writeTo(new StringWriter()).toString();
you can write
  String str = writable.asString();
now.
same exists to get a byte array
  byte[] bytes = writebale.asBytes();
as implmenter of it you must now write into a stream instead of a writer.