[breaking change] add support for java.time
introduce a new interface which defines the defaults for printing
time and dates -> to use it replace Media with BaseMedia which lives
in media-core, that is all. now all printables can print also dates
and times.

decide to split it out from media to allow to override some
configurations without make this methods visible to our consumers.

By default, the date/time is converted to an iso formatted string.
Supported types are:
- offsettime
- ofsetdatetime
and
- localdate
- localtime
- localdatetime
where the offset can be defined by overriding zoneId(), by default the
systemdefault is used